use gRPC client instead of REST/JSON
