+++ cloud/satellite.go
@@ -1,13 +1,9 @@
 package cloud
 
 import (
-	"bytes"
 	"context"
-	"fmt"
-	"net/http"
-	"net/url"
 
-	pipelinesapi "github.com/earthly/cloud-api/pipelines"
+	pb "github.com/earthly/cloud-api/pipelines"
 	"github.com/pkg/errors"
 )
 
@@ -16,30 +12,21 @@ type SatelliteInstance struct {
 	Name     string
 	Org      string
 	Status   string
-	Version  string
 	Platform string
 }
 
 func (c *client) ListSatellites(ctx context.Context, orgID string) ([]SatelliteInstance, error) {
-	url := fmt.Sprintf("/api/v0/satellites?orgId=%s", url.QueryEscape(orgID))
-	status, body, err := c.doCall(ctx, "GET", url, withAuth())
+	resp, err := c.pipelines.ListSatellites(c.withAuth(ctx), &pb.ListSatellitesRequest{
+		OrgId: orgID,
+	})
 	if err != nil {
-		return nil, err
-	}
-	if status != http.StatusOK {
-		return nil, errors.Errorf("failed listing satellites: %s", body)
-	}
-	var resp pipelinesapi.ListSatellitesResponse
-	err = c.jm.Unmarshal(bytes.NewReader([]byte(body)), &resp)
-	if err != nil {
-		return nil, errors.Wrap(err, "failed to unmarshal listTokens response")
+		return nil, errors.Wrap(err, "failed listing satellites")
 	}
 	instances := make([]SatelliteInstance, len(resp.Instances))
 	for i, s := range resp.Instances {
 		instances[i] = SatelliteInstance{
 			Name:     s.Name,
 			Org:      orgID,
-			Version:  s.Version,
 			Platform: s.Platform,
 			Status:   satelliteStatus(s.Status),
 		}
@@ -48,72 +35,58 @@ func (c *client) ListSatellites(ctx context.Context, orgID string) ([]SatelliteI
 }
 
 func (c *client) GetSatellite(ctx context.Context, name, orgID string) (*SatelliteInstance, error) {
-	url := fmt.Sprintf("/api/v0/satellites/%s?orgId=%s", name, url.QueryEscape(orgID))
-	status, body, err := c.doCall(ctx, "GET", url, withAuth())
+	resp, err := c.pipelines.GetSatellite(c.withAuth(ctx), &pb.GetSatelliteRequest{
+		OrgId: orgID,
+		Name:  name,
+	})
 	if err != nil {
-		return nil, err
-	}
-	if status != http.StatusOK {
-		return nil, errors.Errorf("failed listing satellites: %s", body)
-	}
-	var resp pipelinesapi.GetSatelliteResponse
-	err = c.jm.Unmarshal(bytes.NewReader([]byte(body)), &resp)
-	if err != nil {
-		return nil, errors.Wrap(err, "failed to unmarshal listTokens response")
+		return nil, errors.Wrap(err, "failed getting satellite")
 	}
 	return &SatelliteInstance{
 		Name:     name,
 		Org:      orgID,
 		Status:   satelliteStatus(resp.Status),
-		Version:  resp.Version,
 		Platform: resp.Platform,
 	}, nil
 }
 
 func (c *client) DeleteSatellite(ctx context.Context, name, orgID string) error {
-	url := fmt.Sprintf("/api/v0/satellites/%s?orgId=%s", name, url.QueryEscape(orgID))
-	status, body, err := c.doCall(ctx, "DELETE", url,
-		withAuth(), withHeader("Grpc-Timeout", satelliteMgmtTimeout))
+	_, err := c.pipelines.DeleteSatellite(c.withAuth(ctx), &pb.DeleteSatelliteRequest{
+		OrgId: orgID,
+		Name:  name,
+	})
 	if err != nil {
-		return err
-	}
-	if status != http.StatusOK {
-		return errors.Errorf("failed listing satellites: %s", body)
+		return errors.Wrap(err, "failed deleting satellite")
 	}
 	return nil
 }
 
 func (c *client) LaunchSatellite(ctx context.Context, name, orgID string, features []string) error {
-	req := pipelinesapi.LaunchSatelliteRequest{
+	_, err := c.pipelines.LaunchSatellite(c.withAuth(ctx), &pb.LaunchSatelliteRequest{
 		OrgId:        orgID,
 		Name:         name,
 		Platform:     "linux/amd64", // TODO support arm64 as well
 		FeatureFlags: features,
-	}
-	status, body, err := c.doCall(ctx, "POST", "/api/v0/satellites",
-		withAuth(), withHeader("Grpc-Timeout", satelliteMgmtTimeout), withJSONBody(&req))
+	})
 	if err != nil {
-		return err
-	}
-	if status != http.StatusOK {
-		return errors.Errorf("failed launching satellite: %s", body)
+		return errors.Wrap(err, "failed launching satellite")
 	}
 	return nil
 }
 
-func satelliteStatus(status pipelinesapi.SatelliteStatus) string {
+func satelliteStatus(status pb.SatelliteStatus) string {
 	switch status {
-	case pipelinesapi.SatelliteStatus_SATELLITE_STATUS_OPERATIONAL:
+	case pb.SatelliteStatus_SATELLITE_STATUS_OPERATIONAL:
 		return "Operational"
-	case pipelinesapi.SatelliteStatus_SATELLITE_STATUS_SLEEP:
-		return "Sleep"
-	case pipelinesapi.SatelliteStatus_SATELLITE_STATUS_CREATING:
+	case pb.SatelliteStatus_SATELLITE_STATUS_SLEEP:
+		return "Sleeping"
+	case pb.SatelliteStatus_SATELLITE_STATUS_CREATING:
 		return "Creating"
-	case pipelinesapi.SatelliteStatus_SATELLITE_STATUS_FAILED:
+	case pb.SatelliteStatus_SATELLITE_STATUS_FAILED:
 		return "Failed"
-	case pipelinesapi.SatelliteStatus_SATELLITE_STATUS_DESTROYING:
+	case pb.SatelliteStatus_SATELLITE_STATUS_DESTROYING:
 		return "Destroying"
-	case pipelinesapi.SatelliteStatus_SATELLITE_STATUS_OFFLINE:
+	case pb.SatelliteStatus_SATELLITE_STATUS_OFFLINE:
 		return "Offline"
 	default:
 		return "Unknown"
