+++ cmd/earthly/main.go
@@ -106,7 +106,8 @@ type cliFlags struct {
 	disableNewLine            bool
 	secretFile                string
 	secretStdin               bool
-	apiServer                 string
+	cloudHttpAddr             string
+	cloudGrpcAddr             string
 	satelliteAddress          string
 	writePermission           bool
 	registrationPublicKey     string
@@ -270,7 +271,7 @@ func main() {
 		ctxTimeout, cancel := context.WithTimeout(context.Background(), time.Second)
 		defer cancel()
 		displayErrors := app.verbose
-		cloudClient, err := cloud.NewClient(app.apiServer, app.sshAuthSock, app.authToken, app.console.Warnf)
+		cloudClient, err := cloud.NewClient(app.cloudHttpAddr, app.cloudGrpcAddr, app.sshAuthSock, app.authToken, app.console.Warnf)
 		if err != nil && displayErrors {
 			app.console.Warnf("unable to start cloud client: %s", err)
 		} else if err == nil {
